move is_index_alike() to utils
