--- audformat/utils/misc/__init__.py
@@ -1,3 +1,0 @@
-from audformat.core.utils_misc import (
-    is_index_alike,
-)
