--- docs/api-utils-misc.rst
@@ -1,10 +1,0 @@
-audformat.utils.misc
-====================
-
-.. automodule:: audformat.utils.misc
-
-
-is_index_alike
---------------
-
-.. autofunction:: is_index_alike
